[WIP] Healing action throttling
This patch adds a mechanism to throttle the execution of healing
actions. A new `throttling` section will be added to the configuration
file:

```yaml
throttling:
  interval: 1h
```

The new `interval` parameter controls for how long are executed healing
actions remembered. If the same healing action is triggered a second
time within this interval it won't be executed.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>