Use Ansible role to install
Currently the suggested way to install the auto-heal service is to use
the `template.sh` script, which in turn uses the `template.yml`
_OpenShift_ template to create the required objects. This isn't as
flexible as we may need in the future. For example, if we eventually add
an API server we will need to add role bindings to the `kube-system`
namespace. That isn't possible using _OpenShift_ templates because the
`namespace` field is ignored and replaced with the namespace of the
template itself. To address that issue, and to make the suggested way to
install closer to what will be (eventuall) used in the _OpenShift_
installer this patch adds an Ansible role and an ansible playbook to do
the installation. So the installation will now have the following steps:

1. Modify the `install.ini` file to add the minimum set of required
parameters, for example:

```
$ cat >> install.ini <<.
openshift_autoheal_awx_address="https://tower.private/api"
openshift_autoheal_awx_password="EPyZ8Mzw/QeplQ"
.
```

2. Run the `install.sh` script:

```
$ ./install.sh
```

Signed-off-by: Juan Hernandez <jhernand@redhat.com>